fix AI comments

+++ src/main/kotlin/mate/academy/model/Area.kt
@@ -1,5 +1,5 @@
 package mate.academy.model
 
 interface Area {
-    fun calculateArea(): Int
+    fun calculateArea(): Double
 }
